[Refactor] MyBookingRoomCard, MyBookingDetailDialog로 위젯 분리 #11
